Update: changed name of Old Saturation component to Spectrum
