+++ src/components/common/Saturation.js
@@ -1,28 +1,15 @@
 import React, { Component, PureComponent } from 'react'
 import reactCSS from 'reactcss'
-import throttle from 'lodash/throttle'
 import * as saturation from '../../helpers/saturation'
 
 export class Saturation extends (PureComponent || Component) {
-  constructor(props) {
-    super(props)
-
-    this.throttle = throttle((fn, data, e) => {
-      fn(data, e)
-    }, 50)
-  }
-
   componentWillUnmount() {
-    this.throttle.cancel()
     this.unbindEventListeners()
   }
 
   handleChange = (e) => {
-    typeof this.props.onChange === 'function' && this.throttle(
-      this.props.onChange,
-      saturation.calculateChange(e, this.props.hsl, this.container),
-      e,
-    )
+    const change = saturation.calculateChange(e, this.props.direction, this.props.hsl, this.container)
+    change && typeof this.props.onChange === 'function' && this.props.onChange(change, e)
   }
 
   handleMouseDown = (e) => {
@@ -41,73 +28,68 @@ export class Saturation extends (PureComponent || Component) {
   }
 
   render() {
-    const { color, white, black, pointer, circle } = this.props.style || {}
+    const { direction = 'horizontal' } = this.props    
     const styles = reactCSS({
       'default': {
-        color: {
-          absolute: '0px 0px 0px 0px',
-          background: `hsl(${ this.props.hsl.h },100%, 50%)`,
-          borderRadius: this.props.radius,
-        },
-        white: {
+        hue: {
           absolute: '0px 0px 0px 0px',
           borderRadius: this.props.radius,
-        },
-        black: {
-          absolute: '0px 0px 0px 0px',
           boxShadow: this.props.shadow,
+        },
+        container: {
+          padding: '0 2px',
+          position: 'relative',
+          height: '100%',
           borderRadius: this.props.radius,
         },
         pointer: {
           position: 'absolute',
-          top: `${ -(this.props.hsv.v * 100) + 100 }%`,
-          left: `${ this.props.hsv.s * 100 }%`,
-          cursor: 'default',
+          left: `${ this.props.hsl.s }%`,
         },
-        circle: {
+        slider: {
+          marginTop: '1px',
           width: '4px',
-          height: '4px',
-          boxShadow: `0 0 0 1.5px #fff, inset 0 0 1px 1px rgba(0,0,0,.3),
-            0 0 1px 2px rgba(0,0,0,.4)`,
-          borderRadius: '50%',
-          cursor: 'hand',
-          transform: 'translate(-2px, -2px)',
+          borderRadius: '1px',
+          height: '8px',
+          boxShadow: '0 0 2px rgba(0, 0, 0, .6)',
+          background: '#fff',
+          transform: 'translateX(-2px)',
         },
       },
-      'custom': {
-        color,
-        white,
-        black,
-        pointer,
-        circle,
+      'vertical': {
+        pointer: {
+          left: '0px',
+          top: `${ 100 - this.props.hsl.s }%`,
+        },
       },
-    }, { 'custom': !!this.props.style })
+    }, { vertical: direction === 'vertical' })
 
     return (
-      <div
-        style={ styles.color }
-        ref={ container => this.container = container }
-        onMouseDown={ this.handleMouseDown }
-        onTouchMove={ this.handleChange }
-        onTouchStart={ this.handleChange }
-      >
-        <style>{`
-          .saturation-white {
-            background: -webkit-linear-gradient(to right, #fff, rgba(255,255,255,0));
-            background: linear-gradient(to right, #fff, rgba(255,255,255,0));
-          }
-          .saturation-black {
-            background: -webkit-linear-gradient(to top, #000, rgba(0,0,0,0));
-            background: linear-gradient(to top, #000, rgba(0,0,0,0));
-          }
-        `}</style>
-        <div style={ styles.white } className="saturation-white">
-          <div style={ styles.black } className="saturation-black" />
+      <div style={ styles.hue }>
+        <div
+          className={ `saturation-${ direction }` }
+          style={ styles.container }
+          ref={ container => this.container = container }
+          onMouseDown={ this.handleMouseDown }
+          onTouchMove={ this.handleChange }
+          onTouchStart={ this.handleChange }
+        >
+          <style>{ `
+            .saturation-horizontal {
+              background: linear-gradient(to right,hsl(316, 0%, 50%), hsl(${ Math.round(this.props.hsl.h) }, 100%, 50%));
+              background: -webkit-linear-gradient(to right,hsl(316, 0%, 50%), hsl(${ Math.round(this.props.hsl.h) }, 100%, 50%));
+            }
+
+            .saturation-vertical {
+              background: linear-gradient(to top, hsl(316, 0%, 50%), hsl(${ Math.round(this.props.hsl.h) }, 100%, 50%));
+              background: -webkit-linear-gradient(to top,hsl(316, 0%, 50%), hsl(${ Math.round(this.props.hsl.h) }, 100%, 50%));
+            }
+          ` }</style>
           <div style={ styles.pointer }>
             { this.props.pointer ? (
               <this.props.pointer { ...this.props } />
             ) : (
-              <div style={ styles.circle } />
+              <div style={ styles.slider } />
             ) }
           </div>
         </div>
