+++ src/components/common/index.js
@@ -2,7 +2,9 @@ export { default as Alpha } from './Alpha'
 export { default as Checkboard } from './Checkboard'
 export { default as EditableInput } from './EditableInput'
 export { default as Hue } from './Hue'
+export { default as Lightness } from './Lightness'
+// export { default as Hue } from './Hue'
 export { default as Raised } from './Raised'
-export { default as Saturation } from './Saturation'
+export { default as Spectrum } from './Spectrum'
 export { default as ColorWrap } from './ColorWrap'
 export { default as Swatch } from './Swatch'
